refactor(oidc): move OIDC token purge out of Redux into activate COMPASS-9915
Per code review feedback: the purge logic is a pure storage side-effect
with no Redux state involvement, so it belongs in the plugin activate
method alongside the other services rather than as a Redux thunk.
